Refactor #182 remove unnecessary "ORM" sub-namespace of doctrine related files
