+++ src/AppBundle/Doctrine/ProductionFixtureInterface.php
@@ -10,7 +10,7 @@
  * file that was distributed with this source code.
  */
 
-namespace AppBundle\Doctrine\ORM;
+namespace AppBundle\Doctrine;
 
 use Doctrine\Common\DataFixtures\FixtureInterface;
